Move modules in probnum.core to top level
