+++ src/probnum/diffeq/odefiltsmooth/ivp2filter.py
@@ -7,7 +7,7 @@ import numpy as np
 
 from probnum.filtsmooth import ExtendedKalman, UnscentedKalman
 from probnum.filtsmooth.statespace.discrete import DiscreteGaussianModel
-from probnum.core import random_variables as rvs
+from probnum import random_variables as rvs
 
 
 def ivp2ekf0(ivp, prior, evlvar):
