+++ src/probnum/diffeq/odefiltsmooth/prior.py
@@ -14,7 +14,7 @@ from scipy.special import binom  # for Matern
 from scipy.special import factorial  # vectorised factorial for IBM-Q(h)
 
 from probnum.filtsmooth.statespace.continuous import LTISDEModel
-from probnum.core.random_variables import Normal
+from probnum.random_variables import Normal
 
 
 class ODEPrior(LTISDEModel):
