+++ src/probnum/filtsmooth/gaussfiltsmooth/extendedkalman.py
@@ -8,7 +8,7 @@ from probnum.filtsmooth.gaussfiltsmooth.gaussfiltsmooth import (
     GaussFiltSmooth,
     linear_discrete_update,
 )
-from probnum.prob.random_variables import Normal
+from probnum.random_variables import Normal
 from probnum.filtsmooth.statespace import LinearSDEModel, DiscreteGaussianModel
 
 from probnum.filtsmooth.gaussfiltsmooth._utils import is_cont_disc, is_disc_disc
