+++ src/probnum/filtsmooth/gaussfiltsmooth/kalmanposterior.py
@@ -9,7 +9,7 @@ from warnings import warn
 
 import numpy as np
 
-from probnum.core._randomvariablelist import _RandomVariableList
+from probnum._randomvariablelist import _RandomVariableList
 from probnum.filtsmooth.filtsmoothposterior import FiltSmoothPosterior
