+++ src/probnum/filtsmooth/gaussfiltsmooth/unscentedkalman.py
@@ -11,7 +11,7 @@ from probnum.filtsmooth.gaussfiltsmooth.gaussfiltsmooth import (
     GaussFiltSmooth,
     linear_discrete_update,
 )
-from probnum.core.random_variables import Normal
+from probnum.random_variables import Normal
 from probnum.filtsmooth.gaussfiltsmooth.unscentedtransform import UnscentedTransform
 from probnum.filtsmooth.statespace import (
     ContinuousModel,
