+++ src/probnum/filtsmooth/statespace/continuous/linearsdemodel.py
@@ -10,7 +10,7 @@ is a Gauss-Markov process.
 import numpy as np
 import scipy.linalg
 from probnum.filtsmooth.statespace.continuous import continuousmodel
-from probnum.core.random_variables import Normal
+from probnum.random_variables import Normal
 
 __all__ = ["LinearSDEModel", "LTISDEModel"]
