+++ tests/test_diffeq/test_ode/test_ivp.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
 from probnum.diffeq.ode import ivp
-from probnum.core.random_variables import Dirac
+from probnum.random_variables import Dirac
 from tests.testing import NumpyAssertions
 
 TEST_NDIM = 3
