+++ tests/test_diffeq/test_odefiltsmooth/test_ivp2filter.py
@@ -12,7 +12,7 @@ import numpy as np
 
 from probnum.diffeq import IBM, ivp2filter, lotkavolterra
 from probnum.filtsmooth import ExtendedKalman, UnscentedKalman
-from probnum.core.random_variables import Dirac
+from probnum.random_variables import Dirac
 from tests.testing import NumpyAssertions
