+++ tests/test_diffeq/test_odefiltsmooth/test_odefiltsmooth.py
@@ -16,7 +16,7 @@ import numpy as np
 
 from probnum.diffeq.odefiltsmooth import probsolve_ivp
 from probnum.diffeq import ode
-from probnum.core.random_variables import Dirac
+from probnum.random_variables import Dirac
 
 from tests.testing import NumpyAssertions
