+++ tests/test_filtsmooth/test_gaussfiltsmooth/filtsmooth_testcases.py
@@ -12,7 +12,7 @@ from probnum.filtsmooth import (
     generate_cd,
     DiscreteGaussianModel,
 )
-from probnum.core.random_variables import Normal
+from probnum.random_variables import Normal
 from tests.testing import NumpyAssertions
 
 __all__ = [
